requirements.txt
torch
torchvision
transformers
opencv-python-headless
pillow
numpy
streamlit